feat: esc10b post processing
